multiple changes - abandons usage of float32 for float64 - converts console output from string manipulation to template

+++ init.go
@@ -97,8 +97,8 @@ func (i *investments) Set(value string) error {
 		}
 		cmdLnInvestments = append(cmdLnInvestments, investment{
 			Ticker:   inv[0],
-			Quantity: float32(quantity),
-			Price:    float32(price),
+			Quantity: quantity,
+			Price:    price,
 		})
 	}
 	return nil
